MYR-63 Extract gauge wiring + logger/geocoder helpers into wiring.go
Brings cmd/telemetry-server/main.go back under the CLAUDE.md
300-line cap by extracting:

- startPlaintextGauges: registers + runs both rollout health
  gauges (account_token_plaintext_remaining_total +
  vehicle_gps_plaintext_remaining_total).
- newLogger and newGeocoder helpers — pure code-organization
  extractions, no behavior change.

main.go now sits at 285 lines; wiring.go at 293.
